% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flagstaff_Weather.R
\docType{data}
\name{FlagWeatherStation}
\alias{FlagWeatherStation}
\title{Documenting the Flagstaff Pulliam dataset}
\format{
A data frame with 25449 observations with 5 columns.
\describe{
   \item{chr_date}{chr YYYY-MM-DD}
   \item{PRCP}{num Up to two sig figs}
   \item{SNOW}{num Up to two sig figs}
   \item{TMAX}{int in Fahrenheit}
   \item{TMIN}{int in Fahrenheit}
   \item{DATE2}{Date YYYY-MM-DD}
}
}
\usage{
FlagWeatherStation
}
\description{
Here we have data from the 1900s from Flagstaff Pulliam
Weather station. There are many columns, but we are
only concerned with DATE PRCP SNOW TMAX and TMIN
}
\keyword{datasets}
